pyproject.toml
[project]
name = "clawmini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.131.0",
    "html2text>=2025.4.15",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "llama-index>=0.14.15",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
